fix: use CanIRequest and simplify the code.
Updates the policy code to use the CanIRequest type to pass the
kubernetes/can-i host capability paramenters. Furthermore, this commit
also restructure the code to make it more simple.

The policy logic also changed to return a reject as soon as the policy
found a blocked operation. And accept workloads with the pods does not
mount the service account token.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>